Add workflow for documentation deployment and setup new documentation
This commit includes the addition of a GitHub Actions workflow to automatically build, test, and deploy documentation to GitHub Pages. It also introduces new documentation pages, providing an introduction and getting started guide for C++. Configuration for Writerside, including the definition of categories and variables, has also been updated. Lastly, the project license file was renamed for better recognition.